Finishing dashboard and literature page, and api unit test

+++ app/src/main/java/sidev/app/bangkit/capstone/sheltermobile/core/data/local/room/UserDao.kt
@@ -15,5 +15,5 @@ interface UserDao {
     fun getUsers(): List<UserEntity>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun saveUser(data: UserEntity): Int
+    fun saveUser(data: UserEntity): Long
 }
